+++ backend/app_menu_common.go
@@ -3,10 +3,20 @@ package backend
 import (
 	"code.vegaprotocol.io/vegawallet-desktop/app"
 	"github.com/wailsapp/wails/v2/pkg/menu"
+	"github.com/wailsapp/wails/v2/pkg/menu/keys"
 	wailsRuntime "github.com/wailsapp/wails/v2/pkg/runtime"
 )
 
 func (h *Handler) commonAppMenu(appMenu *menu.Menu) {
+	windowMenu := appMenu.AddSubmenu("Window")
+
+	windowMenu.AddText("Minimize", keys.CmdOrCtrl("m"), func(_ *menu.CallbackData) {
+		wailsRuntime.Hide(h.ctx)
+	})
+	windowMenu.AddText("Close", keys.CmdOrCtrl("w"), func(_ *menu.CallbackData) {
+		wailsRuntime.Hide(h.ctx)
+	})
+
 	helpMenu := appMenu.AddSubmenu("Help")
 
 	if softwareDocumentationURL != "" && apiDocumentationURL != "" {